Altered Things BUGGED
Attempted to make the Triangle Summon test be the result of the OnCast function of the Rock Spell.
I'm having trouble referencing variables in the RockSpell Script.

Also, if anyone has any idea how to have the triangle object be summoned without needing it to be in the game work, please do so. <3